linkedin-post-generator/requirements.txt
streamlit==1.32.2
google-generativeai
